Added checks to ensure drone role is present

+++ ai/speech_optimization.py
@@ -3,7 +3,7 @@ import re
 import discord
 from bot_utils import get_id
 from channels import REPETITIONS, ORDERS_REPORTING, ORDERS_COMPLETION
-from roles import SPEECH_OPTIMIZATION, has_role
+from roles import SPEECH_OPTIMIZATION, DRONE, has_role
 from webhook import send_webhook_with_specific_output
 from ai.mantras import Mantra_Handler
 from webhook import get_webhook_for_channel
@@ -100,7 +100,6 @@ CHANNEL_BLACKLIST = [ORDERS_REPORTING, ORDERS_COMPLETION]
 
 
 def get_acceptable_messages(author, channel):
-
     user_id = get_id(author.display_name)
     # Only returns mantra if channels is hexcorp-repetitions; else it returns nothing
     if channel == REPETITIONS:
@@ -113,14 +112,13 @@ def get_acceptable_messages(author, channel):
 
 
 async def print_status_code(message: discord.Message):
-    informative_status_code = informative_status_code_regex.match(
-        message.content)
-    if informative_status_code and not has_role(message.author, SPEECH_OPTIMIZATION):
+    informative_status_code = informative_status_code_regex.match(message.content)
+    if informative_status_code and has_role(message.author, DRONE) and not has_role(message.author, SPEECH_OPTIMIZATION):
         await message.delete()
         return f'{informative_status_code.group(1)} :: Code `{informative_status_code.group(2)}` :: {code_map.get(informative_status_code.group(2), "INVALID CODE")} :: {informative_status_code.group(3)}'
 
     plain_status_code = plain_status_code_regex.match(message.content)
-    if plain_status_code:
+    if plain_status_code and has_role(message.author, DRONE):
         await message.delete()
         return f'{plain_status_code.group(1)} :: Code `{plain_status_code.group(2)}` :: {code_map.get(plain_status_code.group(2), "INVALID CODE")}'
     return False
@@ -139,7 +137,10 @@ async def optimize_speech(message: discord.Message):
         LOGGER.info("Deleting inappropriate message by optimized drone.")
         await message.delete()
         return True
-    elif informative_status_code_message or is_status_code:
+    elif not has_role(message.author, DRONE):
+        LOGGER.info("Ignoring message from non-drone.")
+        return True
+    elif (informative_status_code_message or is_status_code) and has_role(message.author, DRONE):
         LOGGER.info("Optimizing speech code for drone.")
         webhook = await get_webhook_for_channel(message.channel)
         output = await print_status_code(message)
